requirements.txt
click>=6.7
flask>=0.12.2
polyinterface>=2.0.28
pyyaml>=3.12
pychromecast>=7.5.1
requests>=2.18.4
wakeonlan>=1.1.6
ws4py>=0.5.1
